requirements.txt
openai
google-generativeai
flet
tiktoken
# Add other dependencies like tiktoken, etc., as identified    